backend/requirements.txt
fastapi==0.104.1
pydantic[email]==1.10.13
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.22
alembic==1.12.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.1.0
httpx==0.25.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
psycopg2-binary==2.9.9
websockets==12.0
openai==1.54.0
python-dotenv==1.0.0
PyPDF2==3.0.1
